requirements.txt
fastapi
uvicorn
pydantic
aiohttp
requests
python-dotenv